requirement.txt
streamlit
pdfplumber
sentence-transformers
faiss-cpu
google-generativeai
python-dotenv
